lint fix

+++ contracts/Governance/IParameterInfo.sol
@@ -17,8 +17,8 @@ interface IParameterInfo {
     function setPlatformConfig(string memory key, uint256 amount) external returns (bool);
 
     function getPlatformConfig(string memory key) external view returns (uint256);
-    function getUSTNLoanPledgeRateWarningValue() external view returns (uint amount);
-    function getUSTNLoanLiquidationRate() external view  returns (uint amount);
+    function getUSTNLoanPledgeRateWarningValue() external view returns (uint256 amount);
+    function getUSTNLoanLiquidationRate() external view returns (uint256 amount);
 
     function setPriceDiscountConfig(
         uint256 API,
@@ -27,7 +27,9 @@ interface IParameterInfo {
         uint256 discount2,
         uint256 discount3,
         uint256 discount4
-    ) external returns (bool);
+    )
+        external
+        returns (bool);
 
     function getPriceDiscountConfig(uint256 _API, uint256 _sulphur) external view returns (uint256);
 }
